refactor(git): add upstream icon to repo
